pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "harness0"
version = "0.0.3"
description = "Layer 0 of agent reliability — the framework-agnostic harness that makes any AI agent stable."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "seekcontext" },
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "tiktoken>=0.7",
    "httpx>=0.27",
    "aiofiles>=24.0",
]
keywords = [
    "agent",
    "harness",
    "reliability",
    "llm",
    "ai",
    "context-management",
    "tool-governance",
    "security",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/seekcontext/harness0"
Documentation = "https://github.com/seekcontext/harness0/blob/main/docs/USER_MANUAL.md"
Repository = "https://github.com/seekcontext/harness0"
Issues = "https://github.com/seekcontext/harness0/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
